pyproject.toml
[project]
name = "quizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "AliakseiYafremau", email = "aleksey.yafremau@gmail.com" }
]
requires-python = ">=3.11"
dependencies = []

[project.scripts]
quizer = "quizer:main"

[build-system]
requires = ["uv_build>=0.8.11,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "faker>=37.6.0",
    "mypy>=1.17.1",
    "pytest-asyncio>=1.1.0",
    "ruff>=0.12.10",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff.lint]
select = ["INP001", "F401"]
